Convert OpenNI coordinates to canvas coordinates
OpenNI provides skeleton joint coordinates in millimeters with
the origin at the Kinect coordinate.  So, there will be both
positive and negative values.

The HTML Canvas element places the origin in the upper-left-hand corner.

Also make the canvas element the size of the browser window instead of
a fixed size.